branch:      javabot fixed duplicate imports

+++ src/main/java/javabot/commands/Javadoc.java
@@ -24,21 +24,26 @@ public class Javadoc implements Command {
     @Override
     @SuppressWarnings("IOResourceOpenedButNotSafelyClosed")
     public List<Message> execute(final Javabot bot, final BotEvent event, final List<String> args) {
-        final String name = args.get(0);
-        Api api = dao.find(name);
-        if (api != null) {
-            dao.delete(api);
-        }
-        final String urlString = args.get(1);
-        api = new Api(name, urlString);
-        final JavadocParser parser = new JavadocParser();
-        parser.parse(api, new StringWriter() {
-            @Override
-            public void write(final String line) {
-                bot.postMessage(new Message(event.getSender(), event, line));
+        if (args.size() < 2) {
+            bot.postMessage(new Message(event.getChannel(), event, "usage: javadoc <name> <url> [<packages>]"));
+            bot.postMessage(new Message(event.getChannel(), event,
+                "usage: packages are optional.  if not given all packages found will be documented"));
+        } else {
+            final String name = args.remove(0);
+            Api api = dao.find(name);
+            if (api != null) {
+                dao.delete(api);
             }
-        });
+            final String urlString = args.remove(0);
+            api = new Api(name, urlString);
+            final JavadocParser parser = new JavadocParser();
+            parser.parse(api, args, new StringWriter() {
+                @Override
+                public void write(final String line) {
+                    bot.postMessage(new Message(event.getSender(), event, line));
+                }
+            });
+        }
         return Collections.emptyList();
     }
-
 }
